bugfix: can't build with VC19.10 (VS2017)

+++ far/farwinapi.cpp
@@ -1772,7 +1772,7 @@ FAR_SECURITY_DESCRIPTOR GetFileSecurity(const string& Object, SECURITY_INFORMATI
 		DWORD LengthNeeded = 0;
 		if (!::GetFileSecurity(NtObject.data(), RequestedInformation, reinterpret_cast<SECURITY_DESCRIPTOR*>(Buffer), static_cast<DWORD>(Size), &LengthNeeded))
 			return size_t(0);
-		return LengthNeeded <= Size? Size : LengthNeeded;
+		return LengthNeeded <= Size? Size : (size_t)LengthNeeded;
 	},
 	[](size_t ReturnedSize, size_t AllocatedSize)
 	{
